Modular ingredient handlers (#358)
* Factor out "ingredient handler utilities" into a
separate module. This is a prerequisite for factoring out the handlers
themselves.

This also moves the `source` variable into `context`, because `source`
needs to be accessible in both the ingredient handlers and the utils.

* Factor out complex ingredient handlers into their own modules.

This commit factors out the handlers for:
* `data.browser_compatibility`
* `data.specifications`
* `prose.short_description`

into their own modules. It also moves the top-level
"index-handlers.js" to index-handlers/index.js, to sit alongside
the handler modules.

* Add a comment to the 'data.browser_compatibility' handler

* Reinstate blank line separating standard library modules from external libaries